+++ brainbuilder/app/sonata.py
@@ -1,5 +1,6 @@
 """Tools for working with SONATA."""
 # pylint: disable=import-outside-toplevel
+import glob
 import os
 import sys
 
@@ -108,17 +109,27 @@ def network_config(
 
 
 @app.command()
-@click.option("--input-dir", help="Path to the input directory containing the targets files",
-              required=True)
+@click.option("--input-dir", type=REQUIRED_PATH_DIR,
+              help="Path to the input directory containing the targets files")
+@click.option("--target-file", multiple=True, type=REQUIRED_PATH, help="Path to target file")
 @click.option("-c", "--cells-path", help="Path to cells file", required=True)
 @click.option("-o", "--output", help="Path to output the .json file", required=True)
-def node_set_from_targets(input_dir, cells_path, output):
+def node_set_from_targets(input_dir, target_file, cells_path, output):
     """Convert target files into a single node_set.json like file.
 
     Please check 'brainbuilder targets node-sets' also.
     """
     from brainbuilder.utils.sonata.convert import write_node_set_from_targets
-    write_node_set_from_targets(input_dir, output, cells_path)
+
+    target_files = set(target_file)
+
+    if input_dir is not None:
+        target_files = target_files | set(glob.glob(f'{input_dir}/*.target'))
+
+    if target_files:
+        write_node_set_from_targets(target_files, output, cells_path)
+    else:
+        click.secho("No target files specified; output won't exist", fg='red')
 
 
 @app.command()